+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+-   Docstrings
+-   Missing args inside of `edgeseraser.polya`
 ## [0.4.0] - 2022-03-14
 ### Added
 - Support for named vertex labels